Forbid function return value for hybrid functions. Start hybrids refactoring
