+++ src/api/hybrid-functions/compiled-code-symbol.js
