+++ src/compiler/es-next/hybrid-function.js
@@ -2,7 +2,7 @@ import hammerhead from 'testcafe-hammerhead';
 import asyncToGenerator from 'babel-runtime/helpers/asyncToGenerator';
 import { noop, escapeRegExp as escapeRe } from 'lodash';
 import loadBabelLibs from './load-babel-libs';
-import compiledCode from '../../api/common/hybrid/compiled-code-symbol';
+import compiledCode from '../../api/hybrid-functions/compiled-code-symbol';
 import NODE_VER from '../../utils/node-version';
 import { APIError } from '../../errors/runtime';
 import getCallsite from '../../errors/get-callsite';
