+++ test/functional/fixtures/api/es-next/hybrid-function/testcafe-fixtures/hybrid-fn-test.js
@@ -213,21 +213,6 @@ test('Hybrid function without `await`', async () => {
     getLocation();
 });
 
-test('Function return value', async () => {
-    const giveMeAnAnswer = () => new Promise(resolve => window.setTimeout(() => resolve(42), 200));
-
-    const hfn = Hybrid(fn => {
-        const getMyLocation = () => document.location.toString();
-
-        return [fn, getMyLocation];
-    });
-
-    const res = await hfn(giveMeAnAnswer);
-
-    expect(await res[0]()).eql(42);
-    expect(await res[1]()).eql('http://localhost:3000/api/es-next/hybrid-function/pages/index.html');
-});
-
 test('DOM node return value', async () => {
     const getSomeNodes = Hybrid(() => {
         const answer = document.querySelector('.answer');
