+++ test/server/data/hybrid-fn-test-run-tracking/in-promise-then.js
@@ -1,4 +1,4 @@
-import hybridFnTestRunTracker from '../../../../lib/api/common/hybrid/test-run-tracker';
+import hybridFnTestRunTracker from '../../../../lib/api/hybrid-functions/test-run-tracker';
 
 fixture `In Promise.then()`;
