+++ test/server/data/hybrid-fn-test-run-tracking/in-test-function-after-await.js
@@ -1,4 +1,4 @@
-import hybridFnTestRunTracker from '../../../../lib/api/common/hybrid/test-run-tracker';
+import hybridFnTestRunTracker from '../../../../lib/api/hybrid-functions/test-run-tracker';
 
 fixture `In test function after "await"`;
